Added File Logging
At the bottom of the UI, there are controls to log the pressure values as they are displayed to a file.

Log Folder Browser:
The default directory for storing log files is the "Pressure Logs" folder inside the main "Ground Station" directory. Clicking the Browse button allows you to change this.

Log Files:
When recording is enabled, a new log file is created with the naming convention IgniterUILog-YYYYMMDD-HHMMSS.txt. The log files record all events along with a timestamp.